+++ spec/integration/preview/bulk_exports/export_configuration_spec.rb
@@ -31,8 +31,7 @@ describe 'ExportConfiguration' do
           "enabled": true,
           "webhook_url": "",
           "webhook_method": "",
-          "resource_type": "Calls",
-          "email": ""
+          "resource_type": "Calls"
       }
       ]
     ))
@@ -64,7 +63,6 @@ describe 'ExportConfiguration' do
       {
           "url": "https://preview.twilio.com/BulkExports/Exports/Calls/Configuration",
           "enabled": true,
-          "email": "bogus@twilio.com",
           "webhook_url": "",
           "resource_type": "Calls",
           "webhook_method": ""