+++ spec/integration/sync/v1/service_spec.rb
@@ -37,6 +37,7 @@ describe 'Service' do
               "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
           },
           "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "unique_name": "unique_name",
           "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "webhook_url": "http://www.example.com",
           "reachability_webhooks_enabled": false,
@@ -106,6 +107,7 @@ describe 'Service' do
               "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
           },
           "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "unique_name": "unique_name",
           "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "webhook_url": "http://www.example.com",
           "reachability_webhooks_enabled": false,
@@ -184,6 +186,7 @@ describe 'Service' do
                       "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
                   },
                   "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                  "unique_name": "unique_name",
                   "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "webhook_url": "http://www.example.com",
                   "reachability_webhooks_enabled": false,
@@ -229,6 +232,7 @@ describe 'Service' do
               "maps": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Maps"
           },
           "sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "unique_name": "unique_name",
           "url": "https://sync.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "webhook_url": "http://www.example.com",
           "reachability_webhooks_enabled": false,